addNetwork
