+++ src/main.ts
@@ -1,117 +1,14 @@
-export type Address = string;
-export type Base64 = string;
-export type TxnStr = Base64;
-export type SignedTxnStr = Base64;
-
-export interface MultisigMetadata {
-  // Multisig version
-  version: number;
-
-  // Multisig threshold value
-  threshold: number;
-
-  // Multisig Cosigners
-  addrs: Address[];
-}
-
-export interface WalletTransaction {
-  // Base64 encoding of the canonical msgpack encoding of a Transaction
-  txn: TxnStr;
-
-  // Optional authorized address used to sign the transaction when the account is rekeyed
-  authAddr?: Address;
-
-  // [Not Supported] Multisig metadata used to sign the transaction
-  msig?: MultisigMetadata;
-
-  // Optional list of addresses that must sign the transactions
-  signers?: Address[];
-
-  // Optional base64 encoding of the canonical msgpack encoding of a  SignedTxn corresponding to txn, when signers=[]
-  stxn?: SignedTxnStr;
-
-  // [Not Supported] Optional message explaining the reason of the transaction
-  message?: string;
-
-  // [Not Supported] Optional message explaining the reason of this group of transaction.
-  // Field only allowed in the first transaction of a group
-  groupMessage?: string;
-}
-
-export interface SignTxnsOpts {
-  // [Not Supported] Optional message explaining the reason of the group of transactions
-  message?: string;
-  // Name of site requesting signatures
-  _luteSiteName: string;
-}
-
-export class SignTxnsError extends Error {
-  code: number;
-  data?: any;
-
-  constructor(message: string, code: number, data?: any) {
-    super(message);
-    this.name = "SignTxnsError";
-    this.code = code;
-    this.data = data;
-  }
-}
-
-export interface Siwa {
-  domain: string;
-  account_address: string;
-  uri: string;
-  version: string;
-  statement?: string;
-  nonce?: string;
-  "issued-at"?: string;
-  "expiration-time"?: string;
-  "not-before"?: string;
-  "request-id"?: string;
-  chain_id: "283";
-  resources?: string[];
-  type: "ed25519";
-}
-
-export interface SignData {
-  data: string;
-  signer: Uint8Array;
-  domain: string;
-  authenticatorData: Uint8Array;
-  requestId?: string;
-  hdPath?: string;
-  signature?: Uint8Array;
-}
-
-export interface SignDataResponse extends SignData {
-  signature: Uint8Array;
-}
-
-export enum ScopeType {
-  UNKNOWN = -1,
-  AUTH = 1,
-}
-
-export interface SignMetadata {
-  scope: ScopeType;
-  encoding: string;
-}
-
-export class SignDataError extends Error {
-  code: number;
-  data?: any;
-
-  constructor(message: string, code: number, data?: any) {
-    super(message);
-    this.name = "SignDataError";
-    this.code = code;
-    this.data = data;
-  }
-}
-
-interface IWindow extends Window {
-  lute?: Boolean;
-}
+import {
+  AddNetworkError,
+  Address,
+  IWindow,
+  Network,
+  SignDataError,
+  SignDataResponse,
+  SignMetadata,
+  SignTxnsError,
+  WalletTransaction,
+} from "./types";
 
 const left = 100 + window.screenX;
 const top = 100 + window.screenY;
@@ -134,7 +31,7 @@ export default class LuteConnect {
         window.dispatchEvent(
           new CustomEvent("lute-connect", {
             detail: { action: "connect", genesisID },
-          })
+          }),
         );
       } else {
         win = open(`${BASE_URL}/connect`, this.siteName, PARAMS);
@@ -174,7 +71,7 @@ export default class LuteConnect {
         window.dispatchEvent(
           new CustomEvent("lute-connect", {
             detail: { action: "sign", txns },
-          })
+          }),
         );
       } else {
         win = open(`${BASE_URL}/sign`, this.siteName, PARAMS);
@@ -214,7 +111,7 @@ export default class LuteConnect {
         window.dispatchEvent(
           new CustomEvent("lute-connect", {
             detail: { action: "data", data, metadata },
-          })
+          }),
         );
       } else {
         win = open(`${BASE_URL}/auth`, this.siteName, PARAMS);
@@ -242,6 +139,46 @@ export default class LuteConnect {
             reject(new SignDataError("User Rejected Request", 4100));
             break;
         }
+      }
+    });
+  }
+
+  addNetwork(network: Network): Promise<void> {
+    return new Promise(async (resolve, reject) => {
+      const useExt = this.forceWeb ? false : (window as IWindow).lute;
+      let win: any;
+      const type = useExt ? "add-network-response" : "message";
+      window.addEventListener(type, messageHandler);
+      function messageHandler(event: any) {
+        if (!useExt && event.origin !== BASE_URL) return;
+        const detail = event.data || event.detail;
+        if (detail.debug) console.log("[Lute Debug]", detail);
+        switch (detail.action) {
+          case "ready":
+            win?.postMessage({ action: "network", network }, "*");
+            break;
+          case "added":
+            window.removeEventListener(type, messageHandler);
+            resolve(undefined);
+            break;
+          case "error":
+            window.removeEventListener(type, messageHandler);
+            reject(new AddNetworkError(detail.message, detail.code || 4300));
+            break;
+          case "close":
+            window.removeEventListener(type, messageHandler);
+            reject(new AddNetworkError("User Rejected Request", 4100));
+            break;
+        }
+      }
+      if (useExt) {
+        window.dispatchEvent(
+          new CustomEvent("lute-connect", {
+            detail: { action: "network", network },
+          }),
+        );
+      } else {
+        win = open(`${BASE_URL}/network`, this.siteName, PARAMS);
       }
     });
   }
